Tests: Simplify configuration of MIDDLEWARE_CLASSES
In Django 1.9 the default value for `MIDDLEWARE_CLASSES` is a list,
whereas in <= 1.8 it is a tuple. However all versions support lists, so
rather than having to use a try-except, instead always cast to a list.

This will simplify the configuration once we also have to support
Django 1.10's new `MIDDLEWARE` setting.